+++ src/subscription/subscription.entity.ts
@@ -1,6 +1,7 @@
 import { ColumnType, Insertable, Selectable, Updateable } from 'kysely';
 import { Generated } from '../common/kysely-types/generated';
 import { Timestamp } from '../common/kysely-types/timestamp';
+import { PaymentMethodTable } from '../payment/payment.entity';
 
 export interface SubscriptionTable {
 	id: Generated<string>;
@@ -26,29 +27,3 @@ export interface SubscriptionAggregation {
 	subscription: SubscriptionTable;
 	payment_method: PaymentMethodTable;
 }
-
-export type PaymentMethodStatus = 'pending' | 'active';
-
-export interface PaymentMethodTable {
-	id: Generated<string>;
-	user_id: string;
-	payment_method_id: string;
-	status: ColumnType<PaymentMethodStatus, PaymentMethodStatus | undefined, PaymentMethodStatus | undefined>;
-	created_at: Generated<Timestamp>;
-	updated_at: Generated<Timestamp>;
-}
-
-export type PaymentMethod = Selectable<PaymentMethodTable>;
-export type NewPaymentMethod = Insertable<PaymentMethodTable>;
-export type PaymentMethodUpdate = Updateable<PaymentMethodTable>;
-
-export interface PaymentEventTable {
-	id: Generated<string>;
-	user_id: ColumnType<string | null, string | null | undefined>;
-	subscription_id: ColumnType<string | null, string | null | undefined>;
-	event: ColumnType<unknown, unknown>;
-	created_at: Generated<Timestamp>;
-}
-
-export type PaymentEvent = Selectable<PaymentEventTable>;
-export type NewPaymentEvent = Insertable<PaymentEventTable>;
